Add types.go in base package
To improve code modularity and maintainability,
move all check-related type definitions from base.go
to a new file, types.go.
